read_Data changes

+++ read_data.py
@@ -14,19 +14,18 @@ def le_dados():
     caminho_projeto = Path()    
     path_dados_bin = Path(f"{caminho_projeto.absolute()}\\bin\\steam_data_bin.bin")
     
-    #Lê csv completo
-    df = pd.read_csv(f"{caminho_arquivos}\\csv\\steam_data.csv")
-    
-    #Exclui colunas que nao seram tratadas
-    df = df.drop(["all_reviews","pegi","pegi_url"],axis=1)
-   
-    # Exclui as linhas onde a coluna "name" é "-"
-    df = df.drop(df[df['name'] == '-'].index)
-
-    print(df.info())
-
-    with open(path_dados_bin,'wb') as arquivo_bin:
-        pickle.dump(df,arquivo_bin)
+    if not path_dados_bin.is_file():
+        #Lê csv completo
+        df = pd.read_csv(f"{caminho_arquivos}\\csv\\steam_data.csv")
         
-if  __name__ == "__main__":
-    le_dados()
+        #Exclui colunas que nao seram tratadas
+        df = df.drop(["all_reviews","pegi","pegi_url"],axis=1)
+    
+        # Exclui as linhas onde a coluna "name" é "-"
+        df = df.drop(df[df['name'] == '-'].index)
+        rows = df.shape[0]
+        
+        with open(path_dados_bin,'wb') as arquivo_bin:
+            pickle.dump(df,arquivo_bin)
+
+
